database/database.py
import motor.motor_asyncio

from decouple import config

MONGO_DETAILS = config("MONGO_DETAILS")

client = motor.motor_asyncio.AsyncIOMotorClient(MONGO_DETAILS)

database = client.test

weather_collection = database["weather"]
