requirements.txt
uvicorn==0.15.0
pymongo==3.12.0
fastapi==0.68.0
motor==2.5.0
python-decouple==3.4